Add alternate logit postprocessor implementation.
This commit adds a new `Postprocessor` abstract class, refactors the
existing Pandas postprocessing code into the `PandasPostprocessor`
subclass class of that class, and adds an alternate `NumPyPostprocessor`
class. The new postprocessor interface includes a block length parameter
that allows a postprocessor to be tailored for use in a real-time or
non-real-time (e.g. audio file) processing environment. At this point
the `NumPyPostprocessor` class is considerably more efficient for
real-time processing than the `PandasPostprocessor` class, as shown by
the `compare_postprocessor_speeds` script.