Split main source code file into two parts.
The two parts are the detector code, which is now in the file
"detector.py", and the script that runs the detector, now in the file
"run_nighthawk.py". It was clumsy to have a python module called
`nighthawk` within a package with the same name, and that also caused
import problems.

With changes to its "__init__.py" file, the `nighthawk` package now
includes the constants and the `process_files` function that used to be
defined in the `nighthawk` module, and that are now defined in the
`detector` module.